Basic test suite
The Cakefile is the entry point for running tests. There are two suites -
"node" (using mocha from node) and browser (using mocha in a HTML page, which is
served from a small Express application, powered by mocha-phantomjs).
The latter needs some work, but it's essentially a runnable mvstar example
application with an API server backing it, to test the latest features. The
former are basically unit tests. Assertion libraries are provided by chai.

The package.json's test script runs `cake test`, to play with `npm test`.